move OperatorSliceInfo and PidBasis into evolution module
