+++ pineappl_py/src/evolution.rs
@@ -1,7 +1,76 @@
 use numpy::{IntoPyArray, PyArray1};
 use pineappl::evolution::{EvolveInfo, OperatorSliceInfo};
+use pineappl::pids::PidBasis;
 
 use pyo3::prelude::*;
+
+/// PyO3 wrapper to :rustdoc:`pineappl::pids::PidBasis <pids/struct.PidBasis.html>`.
+#[pyclass(name = "PidBasis")]
+#[derive(Clone)]
+pub enum PyPidBasis {
+    Pdg,
+    Evol,
+}
+
+impl From<PyPidBasis> for PidBasis {
+    fn from(basis: PyPidBasis) -> Self {
+        match basis {
+            PyPidBasis::Pdg => Self::Pdg,
+            PyPidBasis::Evol => Self::Evol,
+        }
+    }
+}
+
+/// PyO3 wrapper to :rustdoc:`pineappl::evolution::OperatorSliceInfo <evolution/struct.OperatorSliceInfo.html>`.
+#[pyclass(name = "OperatorSliceInfo")]
+#[derive(Clone)]
+pub struct PyOperatorSliceInfo {
+    pub(crate) info: OperatorSliceInfo,
+}
+
+#[pymethods]
+impl PyOperatorSliceInfo {
+    /// Constructor.
+    ///
+    /// Parameteters
+    /// ------------
+    /// fac0 : float
+    ///     initial factorization scale
+    /// pids0 : list(int)
+    ///     flavors available at the initial scale
+    /// x0 : list(float)
+    ///     x-grid at the initial scale
+    /// fac1 : float
+    ///     evolved final scale
+    /// pids1 : list(int)
+    ///     flavors available at the final scale
+    /// x1 : list(float)
+    ///     x-grid at the final scale
+    /// pid_basis : PyPidBasis
+    ///     flavor basis reprentation at the initial scale
+    #[new]
+    pub fn new(
+        fac0: f64,
+        pids0: Vec<i32>,
+        x0: Vec<f64>,
+        fac1: f64,
+        pids1: Vec<i32>,
+        x1: Vec<f64>,
+        pid_basis: PyPidBasis,
+    ) -> Self {
+        Self {
+            info: OperatorSliceInfo {
+                fac0,
+                pids0,
+                x0,
+                fac1,
+                pids1,
+                x1,
+                pid_basis: pid_basis.into(),
+            },
+        }
+    }
+}
 
 /// PyO3 wrapper to :rustdoc:`pineappl::evolution::EvolveInfo <evolution/struct.EvolveInfo.html>`.
 #[pyclass]
@@ -37,16 +106,10 @@ impl PyEvolveInfo {
     }
 }
 
-/// PyO3 wrapper to :rustdoc:`pineappl::evolution::OperatorSliceInfo <evolution/struct.OperatorSliceInfo.html>`.
-#[pyclass]
-#[repr(transparent)]
-#[derive(Clone)]
-pub struct PyOperatorSliceInfo {
-    pub(crate) slice_info: OperatorSliceInfo,
-}
-
 #[pymodule]
 pub fn evolution(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<PyEvolveInfo>()?;
+    m.add_class::<PyOperatorSliceInfo>()?;
+    m.add_class::<PyPidBasis>()?;
     Ok(())
 }
